pyproject.toml
[project]
name = "lunar"
version = "0.1.0"
description = "A comprehensive trading and backtesting system"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Calvin Wong", email = "calvin.kl.wong@gmail.com" }
]
dependencies = [
    "aiohttp>=3.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "polygon-api-client>=1.12.0",
    "pyarrow>=14.0.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "typing-extensions>=4.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/backtester", "src/trader"]

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
#addopts = "--cov=core --cov=trader --cov-report=term-missing"
addopts = "--cov=core --cov=backtester --cov=trader --cov-report=term-missing"
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "session"